Feat: [BADA-107] map api 상수 적용 #47
