+++ src/features/stores/map/apis/map.ts
@@ -1,0 +1,34 @@
+import { axiosInstance } from '@shared/lib/axios/axiosInstance';
+import { buildQueryParams } from '@utils/buildQueryParams';
+import { ApiResponse } from '@shared/lib/axios/models';
+import {
+  FetchStoresParams,
+  FetchStoreDevicesParams,
+  Store,
+  StoreDevice,
+} from '@features/stores/map/types';
+import { END_POINTS } from '@constants/api';
+import type { AxiosResponse } from 'axios';
+
+/**
+ * 특정 가맹점(storeId)의 장비 목록 조회
+ */
+export const fetchStoreDevices = async (
+  storeId: number,
+  params: FetchStoreDevicesParams,
+): Promise<AxiosResponse<ApiResponse<StoreDevice[]>>> => {
+  const query = buildQueryParams(params);
+  return await axiosInstance.get<ApiResponse<StoreDevice[]>>(
+    END_POINTS.STORES.ALLSTORE(String(storeId), query),
+  );
+};
+
+/**
+ * 전체 가맹점 위치 목록 조회 (조건 필터 포함)
+ */
+export const fetchStores = async (
+  params: FetchStoresParams,
+): Promise<AxiosResponse<ApiResponse<Store[]>>> => {
+  const query = buildQueryParams(params);
+  return await axiosInstance.get<ApiResponse<Store[]>>(END_POINTS.STORES.ALLDEVICE(query));
+};
